requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-redis==5.0.0
djangorestframework==3.12.4
idna==3.3
PyMySQL==1.0.2
pytz==2021.3
qiniu==7.5.0
redis==3.5.3
requests==2.26.0
sqlparse==0.4.2
tencentcloud-sdk-python==3.0.519
typing-extensions==3.10.0.2
urllib3==1.26.7
